migrate colorfont_tables check back to Google Fonts profile
based on feedback from @khaledhosny

(issue #4945)
