+++ Lib/fontbakery/legacy_checkids.py
@@ -16,7 +16,6 @@ renaming_map = {
     "com.google.fonts/check/cjk_not_enough_glyphs":                         "cjk_not_enough_glyphs",
     "com.google.fonts/check/cmap/format_12":                                "cmap/format_12",
     "com.google.fonts/check/color_cpal_brightness":                         "color_cpal_brightness",
-    "com.google.fonts/check/colorfont_tables":                              "colorfont_tables",
     "com.google.fonts/check/contour_count":                                 "contour_count",
     "com.google.fonts/check/family/control_chars":                          "control_chars",
     "com.google.fonts/check/designspace_has_consistent_codepoints":         "designspace_has_consistent_codepoints",
@@ -47,6 +46,7 @@ renaming_map = {
     "com.google.fonts/check/canonical_filename":                            "googlefonts/canonical_filename",
     "com.google.fonts/check/cjk_vertical_metrics":                          "googlefonts/cjk_vertical_metrics",
     "com.google.fonts/check/cjk_vertical_metrics_regressions":              "googlefonts/cjk_vertical_metrics_regressions",
+    "com.google.fonts/check/colorfont_tables":                              "googlefonts/colorfont_tables",
     "com.google.fonts/check/description/broken_links":                      "googlefonts/description/broken_links",
     "com.google.fonts/check/description/eof_linebreak":                     "googlefonts/description/eof_linebreak",
     "com.google.fonts/check/description/family_update":                     "googlefonts/description/family_update",
