+++ Lib/fontbakery/profiles/googlefonts.py
@@ -123,6 +123,7 @@ PROFILE = {
             "googlefonts/canonical_filename",
             "googlefonts/cjk_vertical_metrics",
             "googlefonts/cjk_vertical_metrics_regressions",
+            "googlefonts/colorfont_tables",
             "googlefonts/font_names",
             "googlefonts/fstype",
             "googlefonts/fvar_instances",
