+++ Lib/fontbakery/profiles/universal.py
@@ -31,7 +31,6 @@ PROFILE = {
             "cjk_not_enough_glyphs",
             "cmap/format_12",
             "color_cpal_brightness",
-            "colorfont_tables",
             "contour_count",
             "control_chars",
             "empty_glyph_on_gid1_for_colrv0",
